Install CMake from apt-get if possible

+++ install-deps.sh
@@ -10,8 +10,8 @@ if (echo "$CC" | grep -q "NONE") || (echo "$CXX" | grep -q "NONE"); then
   if uname -a | grep -q "Ubuntu"; then # Ubuntu
     # Install gcc-4.9 and g++-4.9.
     DEBIAN_FRONTEND=noninteractive apt-get -y update
-    DEBIAN_FRONTEND=noninteractive apt-get -y install software-properties-common python-software-properties # for add-apt-repository
-    DEBIAN_FRONTEND=noninteractive add-apt-repository -y ppa:ubuntu-toolchain-r/test # for gcc-4.9 and g++-4.9
+    DEBIAN_FRONTEND=noninteractive apt-get -y install software-properties-common python-software-properties # For add-apt-repository
+    DEBIAN_FRONTEND=noninteractive add-apt-repository -y ppa:ubuntu-toolchain-r/test # For gcc-4.9 and g++-4.9
     DEBIAN_FRONTEND=noninteractive apt-get -y update
     DEBIAN_FRONTEND=noninteractive apt-get -y install gcc-4.9 g++-4.9
   else
@@ -33,14 +33,19 @@ if ! (which cmake && (cmake --version | grep -q "cmake version 3")); then
     # Install via Homebrew.
     sudo -u $SUDO_USER brew update
     sudo -u $SUDO_USER brew install cmake
-  else # Other platforms
-    # Build and install from source.
-    mkdir -p build/cmake-3.5.2
-    tar -xf deps/cmake-3.5.2.tar.gz -C build/cmake-3.5.2 --strip-components=1
-    cd build/cmake-3.5.2 && CC="$CC" CXX="$CXX" ./bootstrap
-    cd build/cmake-3.5.2 && make
-    cd build/cmake-3.5.2 && make install
-    rm -rf build/cmake-3.5.2
+  else
+    if (uname -a | grep -q "Ubuntu") && (DEBIAN_FRONTEND=noninteractive apt-get -Vs install cmake | grep -q "cmake (3\."); then # Ubuntu + apt-get
+      # Install via apt-get.
+      DEBIAN_FRONTEND=noninteractive apt-get -y install cmake
+    else # Other platforms
+      # Build and install from source.
+      mkdir -p build/cmake-3.5.2
+      tar -xf deps/cmake-3.5.2.tar.gz -C build/cmake-3.5.2 --strip-components=1
+      cd build/cmake-3.5.2 && CC="$CC" CXX="$CXX" ./bootstrap
+      cd build/cmake-3.5.2 && make
+      cd build/cmake-3.5.2 && make install
+      rm -rf build/cmake-3.5.2
+    fi
   fi
 fi
